Re-organized the scenario tests.
Re-organized the scenario tests so that we can use them to cover the scenario where latest Net Native tools + last released stable packages are used.

Ref #1048.
