+++ src/System.Private.ServiceModel/tests/Scenarios/Encoding/Encoders/BinaryEncodingTests.4.0.0.cs
