+++ src/System.Private.ServiceModel/tests/Scenarios/Encoding/MessageVersion/MessageVersionTests.4.0.0.cs
