+++ src/System.Private.ServiceModel/tests/Scenarios/Extensibility/WebSockets/Extensibility.WebSockets.Tests.csproj
@@ -17,7 +17,7 @@
     <ProjectGuid>{BAF18A70-9EDA-46E9-80A1-7B5F62215C02}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="WebSocketTests.cs" />
+    <Compile Include="**\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
